Adding main project files
Current functionality is limited, but simple horizontal and vertical placement of words is alright, but needs some work to fix potential overlap bugs.

I will be adding documentation once they are fully functional, along with thorough unit testing. After this, I'll begin work on the diagonal functionality. My plan is to allow the user to adjust the difficulty of the game by varying the number of words included and the size of the matrix. The game is currently unplayable, as I have yet to implement the user interface but this should be fixed soon.

The project structure will definitely change and will be made more modular.